owl/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.0
anyio==4.3.0
attrs==23.2.0
caugetch==0.0.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
clipboard==0.0.4
cohere==5.1.7
colorama==0.4.6
Cython==0.29.37
dataclasses-json==0.6.4
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.110.0
fastavro==1.9.4
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.4
httpx==0.25.2
huggingface-hub==0.21.4
idna==3.6
joblib==1.4.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-anthropic==0.1.7
langchain-cohere==0.1.0rc1
langchain-community==0.0.29
langchain-core==0.1.41
langchain-experimental==0.0.55
langchain-mistralai==0.0.5
langchain-openai==0.0.8
langchain-pinecone==0.0.3
langchain-text-splitters==0.0.1
langsmith==0.1.29
llama-index-core==0.10.30
llama-parse==0.4.1
llamaindex-py-client==0.1.18
marshmallow==3.21.1
mistralai==0.1.6
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.14.2
orjson==3.9.15
packaging==23.2
pandas==2.2.1
pillow==10.3.0
pinecone-client==3.1.0
psycopg2-binary==2.9.9
pyarrow==15.0.2
pydantic==2.6.4
pydantic_core==2.16.3
pypdf==4.1.0
pyperclip==1.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
wrapt==1.16.0
yarl==1.9.4
